+++ modifiers/modifier.py
@@ -62,7 +62,8 @@ class BGE_mod_default(bpy.types.PropertyGroup):
         if self._warning():
             row.alert = True
 
-        row.label(text="{}".format(self.label), icon=self.icon)
+        row.operator('bge.modifier_info', emboss=False, icon=self.icon, text='').modifier_name = self.id
+        row.label(text="{}".format(self.label))
 
         r = row.row(align=True)
         r.enabled = self.active
@@ -72,9 +73,6 @@ class BGE_mod_default(bpy.types.PropertyGroup):
         r = row.row(align=True)
         r.alert = False
         r.alignment = 'RIGHT'
-
-        if bpy.context.preferences.addons[__name__.split('.')[0]].preferences.show_help:
-            r.operator('bge.modifier_info', emboss=False, icon='INFO', text='').modifier_name = self.id
 
         if active_as_x:
             r.prop(self, "active", text="", icon='X', icon_only=True, emboss=False)