+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.3](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_1_3)
+### Fixed
+- UI stopped working when adding the Master Collection as a bundle
+
+
 ## [2.1.2](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_1_2)
 ### Fixed
 - Unity export preset fixed